% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsos-package.R
\docType{package}
\name{dsos-package}
\alias{dsos}
\alias{dsos-package}
\title{dsos: Dataset Shift with Outlier Scores}
\description{
Test for no adverse shift in two-sample comparison when we have a training set, the reference distribution, and a test set. The approach is flexible (extensible) and relies on a robust and powerful test statistic, the weighted AUC. See Kamulete, V. M. (2021) \href{https://arxiv.org/abs/2107.02990}{arXiv:2107.02990} for details. Outlier scores such as trust scores and prediction uncertainty can be used as the basis for comparison for example.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vathymut/dsos}
  \item Report bugs at \url{https://github.com/vathymut/dsos/issues}
}

}
\author{
\strong{Maintainer}: Vathy M. Kamulete \email{vathymut@gmail.com} (\href{https://orcid.org/0000-0002-4451-3743}{ORCID})

Other contributors:
\itemize{
  \item  Royal Bank of Canada (RBC) (Research supported by RBC) [copyright holder]
}

}
\keyword{internal}
