% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-factor.R
\name{pt_from_bayes}
\alias{pt_from_bayes}
\title{Convert Bayesian to Permutation Test.}
\usage{
pt_from_bayes(bayes_test)
}
\arguments{
\item{bayes_test}{A \code{outlier.bayes} object from a Bayesian D-SOS test.}
}
\value{
A named list of class \code{outlier.test} containing:
\itemize{
   \item \code{statistic}: observed WAUC statistic
   \item \code{seq_mct}: sequential Monte Carlo test, when applicable
   \item \code{p_value}: p-value
   \item \code{outlier_scores}: outlier scores from training and test set
}
}
\description{
Convert Bayesian to Permutation Test.
}
\examples{
\donttest{
set.seed(12345)
os_train <- rnorm(n = 3e2)
os_test <- rnorm(n = 3e2)
bayes_test <- bf_from_os(os_train, os_test, use_asymptotic = FALSE)
freq_test <- pt_from_bayes(bayes_test)
freq_test
}

}
\seealso{
Other bayesian-test: 
\code{\link{as_bf}()},
\code{\link{as_pvalue}()},
\code{\link{bf_from_os}()}
}
\concept{bayesian-test}
