% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-factor.R
\name{as_pvalue}
\alias{as_pvalue}
\title{Convert Bayes Factor to P-value}
\usage{
as_pvalue(bf)
}
\arguments{
\item{bf}{Bayes factor.}
}
\value{
P-value (scalar value)
}
\description{
Convert Bayes Factor to P-value
}
\examples{
\donttest{
library(dsos)
pvalue_from_bf <- as_pvalue(bf = 1)
pvalue_from_bf
}

}
\seealso{
[as_bf()] to convert p-value to Bayes factor.

Other bayesian-test: 
\code{\link{as_bf}()},
\code{\link{bf_from_os}()},
\code{\link{pt_from_bayes}()}
}
\concept{bayesian-test}
