% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-factor.R
\name{bf_from_os}
\alias{bf_from_os}
\title{Bayesian Test from Outlier Scores}
\usage{
bf_from_os(
  os_train,
  os_test,
  n_pt = 4000,
  percentile_cutoff = 0.5,
  use_asymptotic = FALSE,
  asymptotic_cutoff = 1/12
)
}
\arguments{
\item{os_train}{Outlier scores in training (reference) set.}

\item{os_test}{Outlier scores in test set.}

\item{n_pt}{The number of permutations.}

\item{percentile_cutoff}{The number of permutations.}

\item{use_asymptotic}{Numeric vector of weights of length}

\item{asymptotic_cutoff}{Numeric vector of weights of length}
}
\value{
A named list of class \code{outlier.bayes} containing:
\itemize{
   \item \code{permuted}: WAUC from permutations, if applicable
   \item \code{posterior}: WAUC from posterior distribution
   \item \code{adverse_threshold}: WAUC threshold for adverse shift
   \item \code{adverse_probability}: probability of adverse shift
   \item \code{bayes_factor}: Bayes factor
   \item \code{outlier_scores}: outlier scores from training and test set
}
}
\description{
Bayesian Test from Outlier Scores
}
\examples{
\donttest{
library(dsos)
set.seed(12345)
os_train <- rnorm(n = 100)
os_test <- rnorm(n = 100)
bayes_pt <- bf_from_os(os_train, os_test)
bayes_pt
# Use the (faster) asymptotic cutoff
bayes_at <- bf_from_os(os_train, os_test, use_asymptotic = TRUE)
bayes_at
# Run in parallel on local cluster
library(future)
future::plan(future::multisession)
bayes_parallel <- bf_from_os(os_train, os_test)
bayes_parallel
}

}
\seealso{
Other bayesian-test: 
\code{\link{as_bf}()},
\code{\link{as_pvalue}()},
\code{\link{pt_from_bayes}()}
}
\concept{bayesian-test}
