% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsos-package.R
\docType{package}
\name{dsos-package}
\alias{dsos}
\alias{dsos-package}
\title{dsos: Dataset Shift with Outlier Scores}
\description{
Test for no adverse shift in two-sample comparison when we have a training set, the reference distribution, and a test set. The approach is flexible and relies on a robust and powerful test statistic, the weighted AUC. Technical details are in Kamulete, V. M. (2021) \href{https://arxiv.org/abs/1908.04000}{arXiv:1908.04000}. Modern notions of outlyingness such as trust scores and prediction uncertainty can be used as the underlying scores for example.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vathymut/dsos}
  \item Report bugs at \url{https://github.com/vathymut/dsos/issues}
}

}
\author{
\strong{Maintainer}: Vathy M. Kamulete \email{vathy.kamulete@rbccm.com} (\href{https://orcid.org/0000-0002-4451-3743}{ORCID})

Other contributors:
\itemize{
  \item  Royal Bank of Canada (RBC) (Research supported and funded by RBC) [copyright holder]
}

}
\keyword{internal}
