% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsos-s3.R
\name{print.outlier.test}
\alias{print.outlier.test}
\title{Print result of test for no adverse shift.}
\usage{
\method{print}{outlier.test}(x, n = 5, ...)
}
\arguments{
\item{x}{A \code{outlier.test} object from a D-SOS test.}

\item{n}{The number of outlier scores to print for each sample.}

\item{...}{Placeholder to be comptatible with S3 `print` generic.}
}
\value{
A \pkg{ggplot2} plot with outlier scores and p-value.
}
\description{
Print result of test for no adverse shift.
}
\examples{
\donttest{
set.seed(12345)
data(iris)
x_train <- iris[1:50, 1:4] # Training sample: Species == 'setosa'
x_test <- iris[51:100, 1:4] # Test sample: Species == 'versicolor'
iris_test <- pt_refit(x_train, x_test, scorer = score_od)
iris_test
}

}
\seealso{
Other s3-method: 
\code{\link{plot.outlier.test}()}
}
\concept{s3-method}
